+++ resources/views/partials/contact_address/map.blade.php
@@ -1,0 +1,10 @@
+@if(count($contactAddresses) && trim(config('contacts.googleMapsKey')) !== '')
+    <div class="card-body">
+        <div class="card-title">
+            Map
+        </div>
+        <div class="card-body">
+            <div id="map"></div>
+        </div>
+    </div>
+@endif